refactor: only one device-service
