+++ app/routes/api.boxes.$deviceId.ts
@@ -9,7 +9,7 @@ import { type Device, type User } from '~/db/schema'
 import { transformDeviceToApiFormat } from '~/lib/device-transform'
 import { getUserFromJwt } from '~/lib/jwt'
 import { StandardResponse } from '~/lib/responses'
-import { deleteDevice } from '~/services/devices-service.server'
+import { deleteDevice } from '~/services/device-service.server'
 
 /**
  * @openapi
